web/requirements.txt
Beaker==1.12.1
bottle==0.12.25
cryptography==40.0.1
et-xmlfile==1.1.0
Jinja2==3.1.2
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
mojimoji==0.0.12
openpyxl==3.1.2
pdfminer.six==20221105
Pillow==9.5.0
pycryptodome==3.17
pymongo==3.10.1
PyPDF2==3.0.1
python-docx==0.8.11
reportlab==3.6.12
requests==2.28.2
selenium==3.141.0
urllib3==1.26.15
